refactor(web): separate concerns and simplify auth providers
